Modified Database/MongoDB/Query.hs:queryRequest to work on GHC 7.2.1.
See the GitHub issue here:
<https://github.com/TonyGen/mongoDB-haskell/issues/5>

The refactored function does not employ the RecordWildCards extension.

To be honest, I'm not sure why that fixed the problem, or even what the problem
truly was. The error message from GHC 7.2.1 (see the GitHub issue) said the
record field qOption was not intialized, so it seemed sensible to refactor the
method and explicity initialize it.